% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datNHP}
\alias{datNHP}
\title{Micro RNA Expression Values for the Non-Human Brain Atlas.}
\format{A data frame with 24085776 rows and 6 variables:
\describe{
 \item{gene}{gene for given sample}
 \item{id_string}{unique identifier for each line}
 \item{value}{expression value for the sample}
 \item{layer}{brain layer the sample originates from}
 \item{subregion}{sample subregion; subregion the sample was taken from}
 \item{age}{sample age}
}}
\source{
\url{"https://github.com/AllenBrainAtlas/DevRhesusLMD"}
}
\usage{
datNHP
}
\description{
A dataset containing the normalized expression values for the Non-Human brain atlas. 
This frame contains the raw expression values as well as location and age information.
}
\keyword{datasets}
