% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dev.expr2}
\alias{dev.expr2}
\title{Micro RNA Expression Values for the Non-Human Brain Atlas.}
\format{A data frame with 143778 rows and 6 variables:
\describe{
 \item{species}{species the sample comes from}
 \item{gene}{gene for given sample}
 \item{log2pcd}{log base 2 pcd value}
 \item{expr}{log base 2 expression value for the sample}
 \item{exprz}{z score for the log base 2 expression}
 \item{escore}{associated escore}
}}
\source{
\url{"https://github.com/AllenBrainAtlas/DevRhesusLMD"}
}
\usage{
dev.expr2
}
\description{
A dataset containing the normalized expression values for the Non-Human brain atlas. 
Contains samples for mouse, rat, human, and macaque. This data set is used to create
times series plots across species for gene expression.
}
\keyword{datasets}
