% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprl2}
\alias{exprl2}
\title{Micro RNA Expression Values for the Non-Human Brain Atlas.}
\format{A data frame with 8410116 rows and 6 variables:
\describe{
 \item{gene}{gene measured }
 \item{sample_id}{Allen Institute sample ID}
 \item{brain_structure}{brain structure acronym}
 \item{age}{Specimen age}
 \item{area}{Brain area the sample originates from}
 \item{layer}{Brain layer the sample originates from}
}}
\source{
\url{"https://github.com/AllenBrainAtlas/DevRhesusLMD"}
}
\usage{
exprl2
}
\description{
A dataset containing the normalized expression values for the Non-Human brain atlas. This
data set contains the nesecary information for plotting developmental time series across
a subset or cortex regions.
}
\keyword{datasets}
