% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probes}
\alias{probes}
\title{Metadata for probes used in the Non-human brain atlas}
\format{A data frame with 12411 rows and 10 variables:}
\source{
\url{"https://github.com/AllenBrainAtlas/DevRhesusLMD"}
}
\usage{
probes
}
\description{
Metadata for probes used in the Non-human brain atlas
}
\keyword{datasets}
